==> script_21_base_gpp.py <==
import matplotlib.pylab as plt
import numpy as np

x0 = [10.76106, 5185.45, 11.11234, 10.00857]
x1 = [10.76135, 5154.01, 11.30035, 10.34981]
x2 = [10.75842, 5139.34, 10.92557, 9.895722]
x3 = [10.75549, 5105.70, 10.98399, 9.940643]
x4 = [10.75581, 5106.43, 11.18305, 10.17857]
x5 = [10.75815, 5115.36, 11.10981, 10.17914]
x6 = [10.76114, 5334.14, 11.18026, 10.33888]
x7 = [10.75659, 5105.16, 11.21217, 9.599007] # This is the weird one
x8 = [10.76217, 5150.87, 11.21821, 10.33835]
x9 = [10.76366, 5375.13, 10.82854, 9.870513]

x = np.array([x1,x2,x3,x4,x5,x6,x7,x8,x9])

print(x.shape)

fig, (ax1, ax2, ax3) = plt.subplots(3,1)
ax1.hist(x[:,0]*100, bins = 50)
ax2.hist(x[:,2]*100, bins = 50)
ax3.hist(x[:,3]*100, bins = 50)

plt.savefig('scratch_3.png')
